fixing amount validation

+++ app/Http/Requests/SimpleInscriptionRequest.php
@@ -25,7 +25,7 @@ class SimpleInscriptionRequest extends FormRequest
             'extra.place' => [ 
                 Rule::in(['montevideo', 'interior'])
             ],
-            'amount'=>'float',
+            'amount'=>'numeric',
             'institution_name'=>'string|max:255',
             'institution_type'=>'string|max:255',
             'type' =>  new Enum(InscriptionTypeEnum::class)
